Update Swapping numbers.c

+++ Swapping numbers.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 int main()
 {
-int num_1,num_2;
+float num_1,num_2;
 printf("Enter the first number: ");
-scanf("%d",&num_1);
+scanf("%f",&num_1);
 printf("Enter the second number: ");
-scanf("%d",&num_2);
-printf("originally num1= %d and num2= %d\n",num_1,num_2);
+scanf("%f",&num_2);
+printf("originally num1= %f and num2= %f\n",num_1,num_2);
 num_1= num_1 + num_2;
 num_2= num_1 - num_2;
 num_1= num_1 - num_2;
-printf("After swapping  num1= %d and num2= %d",num_1,num_2);
+printf("After swapping  num1= %f and num2= %f",num_1,num_2);
 }
